+++ Sources/SKTiledScene.swift
@@ -18,7 +18,7 @@ import SpriteKit
  - parameter cameraNode: `SKTiledSceneCamera!` scene camera node.
  - parameter tilemap:    `SKTilemap!` tile map node.
  */
-public protocol SKTiledSceneDelegate: SKTilemapDelegate {
+public protocol SKTiledSceneDelegate {
     /** 
      World container node. All Tiled assets are parented to this node.
     */
@@ -37,7 +37,7 @@ public protocol SKTiledSceneDelegate: SKTilemapDelegate {
  - parameter cameraNode: `SKTiledSceneCamera!` scene camera node.
  - parameter tilemap:    `SKTilemap!` tile map node.
  */
-open class SKTiledScene: SKScene, SKPhysicsContactDelegate, SKTiledSceneDelegate {
+open class SKTiledScene: SKScene, SKPhysicsContactDelegate, SKTiledSceneDelegate, SKTilemapDelegate {
     
     /// World container node.
     open var worldNode: SKNode!
@@ -82,8 +82,8 @@ open class SKTiledScene: SKScene, SKPhysicsContactDelegate, SKTiledSceneDelegate
     
     override open func sceneDidLoad() {
         setupWorld()
-    }    
-
+    }
+    
     override open func didMove(to view: SKView) {
         guard let worldNode = worldNode else { return }
         
@@ -96,7 +96,7 @@ open class SKTiledScene: SKScene, SKPhysicsContactDelegate, SKTiledSceneDelegate
         // load the current tmx file name
         guard let tmxFilename = tmxFilename else { return }
         
-        if let tilemap = loadTilemap(fromFile: tmxFilename) {
+        if let tilemap = load(fromFile: tmxFilename ) {
             // add the tilemap to the world container node.
             worldNode.addChild(tilemap)
             self.tilemap = tilemap
@@ -116,7 +116,6 @@ open class SKTiledScene: SKScene, SKPhysicsContactDelegate, SKTiledSceneDelegate
             }
         }
     }
-    
     
     // MARK: - Setup
     
@@ -146,11 +145,11 @@ open class SKTiledScene: SKScene, SKPhysicsContactDelegate, SKTiledSceneDelegate
     /**
      Load a named TMX file.
      
-     - parameter fileNamed: `String` TMX file name.
+     - parameter filename:  `String` TMX file name.
      - returns: `SKTilemap?` tile map node.
      */
-    open func loadTilemap(fromFile filename: String, completion: (() -> ())? = nil) -> SKTilemap? {
-        if let tilemapNode = SKTilemap.load(fromFile: filename, delegate: self, completion: completion) {
+    open func load(fromFile filename: String) -> SKTilemap? {
+        if let tilemapNode = SKTilemap.load(fromFile: filename, delegate: self) {
             if (tilemapNode.backgroundColor != nil) {
                 self.backgroundColor = tilemapNode.backgroundColor!
             }
@@ -160,12 +159,23 @@ open class SKTiledScene: SKScene, SKPhysicsContactDelegate, SKTiledSceneDelegate
     }
     
     // MARK: - Callbacks
-    /**
-     Callback for post-processing the rendered tilemap.
-     
-     - parameter tilemap: `SKTilemap` tile map node.
-     */
-    open func didRenderMap(_ tilemap: SKTilemap, completion: (() -> ())? = nil) {
-        // placeholder for users to develop their own callbacks
+    open func didBeginParsing(_ tilemap: SKTilemap) {
+        // Called when tilemap is instantiated.
+    }
+    
+    open func didAddTileset(_ tileset: SKTileset) {
+        // Called when a tileset has been added.
+    }
+    
+    open func didAddLayer(_ layer: TiledLayerObject) {
+        // Called when a layer has been added.
+    }
+    
+    open func didReadMap(_ tilemap: SKTilemap) {
+        // Called before layers are rendered.
+    }
+    
+    open func didRenderMap(_ tilemap: SKTilemap) {
+        // Called after layers are rendered. Perform any post-processing here.
     }
 }