+++ Sources/SKTilemapParser.swift
@@ -49,6 +49,8 @@ The `SKTilemapParser` is a custom `XMLParserDelegate` parser for reading Tiled T
  */
 open class SKTilemapParser: NSObject, XMLParserDelegate {
     
+    //let queue: DispatchQueue = DispatchQueue.global(qos: .userInteractive)
+    
     open var fileNames: [String] = []                               // list of files to read
     open var currentFileName: String!
     
@@ -77,15 +79,17 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
     /**
      Load a TMX file and parse it.
      
-     - parameter fileNamed: `String` file base name - TMX/TSX not required.     
+     - parameter filename:   `String` Tiled file name (does not need TMX extension).
+     - parameter delegate:   `SKTilemapDelegate?` optional tilemap delegate instance.
      - returns: `SKTilemap?` tiled map node.
      */
-    open func load(fromFile filename: String, delegate: SKTilemapDelegate? = nil, completion: (() -> ())? = nil) -> SKTilemap? {
+    open func load(fromFile filename: String, delegate: SKTilemapDelegate? = nil) -> SKTilemap? {
         guard let targetFile = getBundledFile(named: filename) else {
             print("[SKTilemapParser]: unable to locate file: \"\(filename)\"")
             return nil
         }
         
+        // set the delegate property
         mapDelegate = delegate
         timer = Date()
         fileNames.append(targetFile)
@@ -134,7 +138,7 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
         externalTilesets = [:]
         
         // render and complete
-        didFinishParsing(completion: completion)
+        didFinishParsing()
         
         return tilemap
     }
@@ -160,26 +164,43 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
         return nil
     }
 
-    // MARK: - Callbacks
+    // MARK: - Post-Processing
+    
+    /**
+     Post-process called when tilemap is finished parsing.     
+     */
+    fileprivate func didFinishParsing() {
+        guard let tilemap = tilemap else { return }
+        
+        let queue = DispatchQueue.main
+        let parseGroup = DispatchGroup()
+        
+        // pre-processing callback
+        queue.async(group: parseGroup) {
+            if self.mapDelegate != nil { self.mapDelegate!.didReadMap(tilemap) }
+        }
+        
+        // start rendering layers when queue is complete.
+        parseGroup.notify(queue: DispatchQueue.main) {
+            self.didBeginRendering(tilemap)
+        }
+    }
     
     /**
      Post-process to render each layer.
      
+     - parameter tilemap:  `SKTilemap`    tile map node.
      - parameter duration: `TimeInterval` fade-in time for each layer.
      */
-    fileprivate func didFinishParsing(duration: TimeInterval=0.075, completion: (() -> ())? = nil)  {
-        guard let tilemap = tilemap else { return }
+    fileprivate func didBeginRendering(_ tilemap: SKTilemap, duration: TimeInterval=0.05)  {
+        // worker queue & group
+        let queue = DispatchQueue.global(qos: .userInteractive)
+        let renderGroup = DispatchGroup()
         
-        // worker queue
-        let queue = DispatchQueue.global(qos: .userInitiated )
+        queue.async(group: renderGroup) {
         
-        // create a group for each tile layer
-        let tileGroup = DispatchGroup()
-        
-        queue.async(group: tileGroup){
-            
             for layer in tilemap.allLayers() {
-                
+
                 // render object groups
                 if let objectGroup = layer as? SKObjectGroup {
                     objectGroup.drawObjects()
@@ -195,13 +216,12 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
                 
                 // render tile layers
                 if let tileLayer = layer as? SKTileLayer {
-                    
                     guard let tileData = self.data[tileLayer.uuid] else { continue }
                     // add the layer data and completion handler
                     let _ = tileLayer.setLayerData(tileData, completion: { (_ layer: SKTileLayer) -> Void in
-                        // run the tilemap completion handler
+                        // run the tile layer completion callback
                         tileLayer.didFinishRendering(duration: duration)
-                        
+
                     })
                 
                     // report errors
@@ -213,12 +233,13 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
                 }
             }
         }
-        
+
         // reset the data property when all layers are rendered & run a callback on the tilemap node
-        tileGroup.notify(queue: DispatchQueue.main) {
+        renderGroup.notify(queue: DispatchQueue.main) {
             self.data = [:]
-            self.tilemap.didFinishParsing(timeStarted: self.timer, completion: completion)
+            self.tilemap.didFinishRendering(timeStarted: self.timer)
         }
+        
     }
     
     // MARK: - XMLParserDelegate
@@ -251,8 +272,9 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
             let currentBasename = currentFileName.components(separatedBy: ".").first!
             self.tilemap.filename = currentBasename
             self.tilemap.name = currentBasename
+            
             // run setup functions on tilemap
-            self.tilemap.didBeginParsing()
+            self.mapDelegate?.didBeginParsing(tilemap)
             lastElement = tilemap
         }
         
@@ -275,6 +297,9 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
                     externalTilesets[source] = tileset
                     self.tilemap.addTileset(tileset)
                     lastElement = tileset
+                    
+                    // delegate callback
+                    if mapDelegate != nil { mapDelegate!.didAddTileset(tileset) }
 
                     // set this to nil, just in case we're looking for a collections tileset.
                     currentID = nil
@@ -312,6 +337,9 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
                     self.tilemap.addTileset(tileset)
                     lastElement = tileset
 
+                    // delegate callback
+                    if mapDelegate != nil { mapDelegate!.didAddTileset(tileset) }
+                    
                     // set this to nil, just in case we're looking for a collections tileset.
                     currentID = nil
                 }
@@ -331,7 +359,7 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
         if elementName == "property" {
             guard let name = attributeDict["name"] else { parser.abortParsing(); return }
             guard let value = attributeDict["value"] else { parser.abortParsing(); return }
-            //guard let value = attributeDict["type"] else { parser.abortParsing(); return }
+            //guard let propertyType = attributeDict["type"] else { parser.abortParsing(); return }
             
             // stash properties
             properties[name] = value            
@@ -349,6 +377,10 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
             }
             
             self.tilemap!.addLayer(layer)
+            
+            // delegate callback
+            if mapDelegate != nil { mapDelegate!.didAddLayer(layer) }
+            
             lastElement = layer
         }
         
@@ -364,6 +396,10 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
             }
             
             self.tilemap!.addLayer(objectsGroup)
+            
+            // delegate callback
+            if mapDelegate != nil { mapDelegate!.didAddLayer(objectsGroup) }
+            
             lastElement = objectsGroup
         }
         
@@ -377,6 +413,10 @@ open class SKTilemapParser: NSObject, XMLParserDelegate {
             }
             
             self.tilemap!.addLayer(imageLayer)
+            
+            // delegate callback
+            if mapDelegate != nil { mapDelegate!.didAddLayer(imageLayer) }
+            
             lastElement = imageLayer
         }